Move more Inventory events to API.
