+++ src/main/java/org/spout/vanilla/plugin/component/substance/material/Anvil.java
@@ -30,16 +30,15 @@ import org.spout.api.Spout;
 import org.spout.api.entity.Player;
 import org.spout.api.inventory.Inventory;
 
-import org.spout.vanilla.api.component.substance.material.ViewedBlockComponent;
-import org.spout.vanilla.api.inventory.Container;
+import org.spout.vanilla.api.component.substance.material.AnvilComponent;
 
 import org.spout.vanilla.plugin.component.inventory.WindowHolder;
-import org.spout.vanilla.plugin.event.inventory.AnvilCloseEvent;
-import org.spout.vanilla.plugin.event.inventory.AnvilOpenEvent;
+import org.spout.vanilla.api.event.inventory.AnvilCloseEvent;
+import org.spout.vanilla.api.event.inventory.AnvilOpenEvent;
 import org.spout.vanilla.plugin.inventory.block.AnvilInventory;
 import org.spout.vanilla.plugin.inventory.window.block.AnvilWindow;
 
-public class Anvil extends ViewedBlockComponent implements Container {
+public class Anvil extends AnvilComponent {
 	private final AnvilInventory inventory = new AnvilInventory();
 
 	@Override
