+++ src/main/java/org/spout/vanilla/plugin/component/substance/material/BrewingStand.java
@@ -29,16 +29,15 @@ package org.spout.vanilla.plugin.component.substance.material;
 import org.spout.api.Spout;
 import org.spout.api.entity.Player;
 
-import org.spout.vanilla.api.component.substance.material.ViewedBlockComponent;
-import org.spout.vanilla.api.inventory.Container;
+import org.spout.vanilla.api.component.substance.material.BrewingStandComponent;
 
 import org.spout.vanilla.plugin.component.inventory.WindowHolder;
-import org.spout.vanilla.plugin.event.inventory.BrewingStandCloseEvent;
-import org.spout.vanilla.plugin.event.inventory.BrewingStandOpenEvent;
+import org.spout.vanilla.api.event.inventory.BrewingStandCloseEvent;
+import org.spout.vanilla.api.event.inventory.BrewingStandOpenEvent;
 import org.spout.vanilla.plugin.inventory.block.BrewingStandInventory;
 import org.spout.vanilla.plugin.inventory.window.block.BrewingStandWindow;
 
-public class BrewingStand extends ViewedBlockComponent implements Container {
+public class BrewingStand extends BrewingStandComponent {
 	private final BrewingStandInventory inventory = new BrewingStandInventory();
 
 	@Override
