+++ src/main/java/org/spout/vanilla/plugin/protocol/VanillaNetworkSynchronizer.java
@@ -95,12 +95,12 @@ import org.spout.vanilla.plugin.event.entity.network.EntityEffectEvent;
 import org.spout.vanilla.plugin.event.entity.network.EntityRemoveEffectEvent;
 import org.spout.vanilla.api.event.item.MapItemUpdateEvent;
 import org.spout.vanilla.plugin.event.player.network.PlayerAbilityUpdateEvent;
-import org.spout.vanilla.plugin.event.player.network.PlayerBedEvent;
-import org.spout.vanilla.plugin.event.player.network.PlayerGameStateEvent;
-import org.spout.vanilla.plugin.event.player.network.PlayerHealthEvent;
-import org.spout.vanilla.plugin.event.player.network.PlayerListEvent;
-import org.spout.vanilla.plugin.event.player.network.PlayerPingEvent;
-import org.spout.vanilla.plugin.event.player.network.PlayerSelectedSlotChangeEvent;
+import org.spout.vanilla.api.event.player.network.PlayerBedEvent;
+import org.spout.vanilla.api.event.player.network.PlayerGameStateEvent;
+import org.spout.vanilla.api.event.player.network.PlayerHealthEvent;
+import org.spout.vanilla.api.event.player.network.PlayerListEvent;
+import org.spout.vanilla.api.event.player.network.PlayerPingEvent;
+import org.spout.vanilla.api.event.player.network.PlayerSelectedSlotChangeEvent;
 import org.spout.vanilla.plugin.event.window.WindowCloseEvent;
 import org.spout.vanilla.plugin.event.window.WindowItemsEvent;
 import org.spout.vanilla.plugin.event.window.WindowOpenEvent;
