+++ src/main/java/org/spout/vanilla/plugin/protocol/handler/player/PlayerHeldItemChangeHandler.java
@@ -33,7 +33,7 @@ import org.spout.api.protocol.Session;
 
 import org.spout.vanilla.plugin.component.living.neutral.Human;
 import org.spout.vanilla.api.event.player.PlayerHeldItemChangeEvent;
-import org.spout.vanilla.plugin.event.player.network.PlayerSelectedSlotChangeEvent;
+import org.spout.vanilla.api.event.player.network.PlayerSelectedSlotChangeEvent;
 import org.spout.vanilla.plugin.inventory.player.PlayerQuickbar;
 import org.spout.vanilla.plugin.protocol.msg.player.PlayerHeldItemChangeMessage;
 import org.spout.vanilla.plugin.util.PlayerUtil;
